Move code in job_scheduler.py . Rename get_site_id in get_repo_id.
