+++ src/vorta/tray_menu.py
@@ -3,7 +3,7 @@ from PyQt5.QtWidgets import QMenu, QSystemTrayIcon
 from PyQt5.QtGui import QIcon
 
 from vorta.models import BackupProfileModel
-from vorta.scheduler import JobsManager
+from vorta.borg.job_scheduler import JobsManager
 from vorta.utils import get_asset
